http_server: extract window rename handler into its own submodule
