+++ daemon/http_server/src/handlers.rs
